+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.handlerAdapter.HandlerAdapter;
@@ -6,6 +6,7 @@ import com.interface21.webmvc.servlet.mvc.handlerAdapter.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mvc.handlerAdapter.HandlerExecutionHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.handlerMapping.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.handlerMapping.HandlerMappingRegistry;
+import jakarta.servlet.ServletContext;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -19,8 +20,10 @@ public class DispatcherServlet extends HttpServlet {
 
     private final HandlerMappingRegistry handlerMappingRegistry;
     private final HandlerAdapterRegistry handlerAdapterRegistry;
+    private final String basePackage;
 
-    public DispatcherServlet() {
+    public DispatcherServlet(ServletContext servletContext) {
+        this.basePackage = (String) servletContext.getAttribute("basePackage");
         this.handlerMappingRegistry = new HandlerMappingRegistry();
         this.handlerAdapterRegistry = new HandlerAdapterRegistry();
     }
@@ -32,7 +35,7 @@ public class DispatcherServlet extends HttpServlet {
     }
 
     private void initHandlerMappingRegistry() {
-        AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(Application.class);
+        AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(basePackage);
         annotationHandlerMapping.initialize();
 
         handlerMappingRegistry.addHandlerMapping(annotationHandlerMapping);
