+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/handlerAdapter/HandlerAdapterRegistryTest.java
@@ -3,14 +3,12 @@ package com.interface21.webmvc.servlet.mvc.handlerAdapter;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import java.lang.reflect.Method;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
-import samples.TestLegacyController;
 
 class HandlerAdapterRegistryTest {
 
@@ -22,20 +20,8 @@ class HandlerAdapterRegistryTest {
 
     @BeforeEach
     void setUp() {
-        ControllerHandlerAdapter controllerHandlerAdapter = new ControllerHandlerAdapter();
         HandlerExecutionHandlerAdapter executionHandlerAdapter = new HandlerExecutionHandlerAdapter();
-        handlerAdapterRegistry.addHandlerAdapter(controllerHandlerAdapter);
         handlerAdapterRegistry.addHandlerAdapter(executionHandlerAdapter);
-    }
-
-    @DisplayName("컨트롤러를 인자로 받으면 ControllerHandlerAdapter를 반환한다.")
-    @Test
-    void getControllerHandlerAdapter() {
-        Controller controller = new TestLegacyController();
-        HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(controller);
-
-        assertThat(handlerAdapter)
-                .isInstanceOf(ControllerHandlerAdapter.class);
     }
 
     @DisplayName("HandlerExecution을 인자로 받으면 HandlerExecutionHandlerAdapter를 반환한다.")
