+++ mvc/src/test/java/samples/TestController.java
@@ -38,4 +38,12 @@ public class TestController {
         modelAndView.addObject("id", request.getAttribute("id"));
         return modelAndView;
     }
+
+    @RequestMapping(value = "/get-hi-jsp")
+    public ModelAndView getHiJsp(final HttpServletRequest request, final HttpServletResponse response) {
+        log.info("test controller without method");
+        final var modelAndView = new ModelAndView(new JspView("hi.jsp"));
+        modelAndView.addObject("id", request.getAttribute("id"));
+        return modelAndView;
+    }
 }
