--- mvc/src/test/java/samples/TestHandlerMapping.java
@@ -1,23 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.handlerMapping.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-
-public class TestHandlerMapping implements HandlerMapping {
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        controllers.put("/test-legacy", new TestLegacyController());
-    }
-
-    @Override
-    public Controller getHandler(final HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        return controllers.get(requestURI);
-    }
-}
