--- mvc/src/test/java/samples/TestLegacyController.java
@@ -1,12 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class TestLegacyController implements Controller {
-
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return "path";
-    }
-}
